% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_asterisks.R, R/signif.R
\name{add_asterisks}
\alias{add_asterisks}
\title{signif}
\usage{
add_asterisks(pvector)

add_asterisks(pvector)
}
\arguments{
\item{vector}{vector of p.values}

\item{vector}{vector of p.values}
}
\description{
adds the asterisks for a vector of p values

adds the asterisks for a vector of p values
}
\examples{
\dontrun{
hello()
}
\dontrun{
hello()
}
}
